Small Update to add /all route
New localhost:3000/all returns everything in a single object without posting it to the server. UpdateServer has a returnOnly parameter that defaults to false; calling it with arguments ("", true) will only return, not post. The returnOnly parameter  is optional so existing calls will continue to work as they did before.